refactor: split utils module
